+++ syn/quartus/mts_processor_static.f
@@ -1,0 +1,3 @@
+../../mts_processor.vhd
+dual_port_rom_blackbox.sv
+mts_processor_syn_top.vhd